+++ private/buf/cmd/buf/command/alpha/protoc/plugin.go
@@ -59,6 +59,16 @@ func executePlugin(
 		storageosProvider,
 		runner,
 	)
+	requests, err := bufimage.ImagesToCodeGeneratorRequests(
+		images,
+		strings.Join(pluginInfo.Opt, ","),
+		bufpluginexec.DefaultVersion,
+		false,
+		false,
+	)
+	if err != nil {
+		return nil, err
+	}
 	var options []bufpluginexec.GenerateOption
 	if pluginInfo.Path != "" {
 		options = append(options, bufpluginexec.GenerateWithPluginPath(pluginInfo.Path))
@@ -67,13 +77,7 @@ func executePlugin(
 		ctx,
 		container,
 		pluginName,
-		bufimage.ImagesToCodeGeneratorRequests(
-			images,
-			strings.Join(pluginInfo.Opt, ","),
-			bufpluginexec.DefaultVersion,
-			false,
-			false,
-		),
+		requests,
 		options...,
 	)
 	if err != nil {